tui: spin off shim_support
